LPS-63674 Move every configurator to the proper location given the new dir structure for XStreamConfigurator
