+++ modules/apps/web-experience-management/journal/journal-service/src/main/java/com/liferay/journal/xstream/configurator/JournalXStreamConfigurator.java
@@ -1,0 +1,63 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.journal.xstream.configurator;
+
+import com.liferay.exportimport.kernel.xstream.XStreamAlias;
+import com.liferay.exportimport.kernel.xstream.XStreamConverter;
+import com.liferay.exportimport.kernel.xstream.XStreamType;
+import com.liferay.journal.model.impl.JournalArticleImpl;
+import com.liferay.journal.model.impl.JournalFeedImpl;
+import com.liferay.journal.model.impl.JournalFolderImpl;
+import com.liferay.portal.kernel.util.ListUtil;
+import com.liferay.xstream.configurator.XStreamConfigurator;
+
+import java.util.List;
+
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Mate Thurzo
+ */
+@Component(immediate = true, service = XStreamConfigurator.class)
+public class JournalXStreamConfigurator implements XStreamConfigurator {
+
+	@Override
+	public List<XStreamType> getAllowedXStreamTypes() {
+		return null;
+	}
+
+	@Override
+	public List<XStreamAlias> getXStreamAliases() {
+		return ListUtil.toList(_xStreamAliases);
+	}
+
+	@Override
+	public List<XStreamConverter> getXStreamConverters() {
+		return null;
+	}
+
+	@Activate
+	protected void activate() {
+		_xStreamAliases = new XStreamAlias[] {
+			new XStreamAlias(JournalArticleImpl.class, "JournalArticle"),
+			new XStreamAlias(JournalFeedImpl.class, "JournalFeed"),
+			new XStreamAlias(JournalFolderImpl.class, "JournalFolder")
+		};
+	}
+
+	private XStreamAlias[] _xStreamAliases;
+
+}
